# API_BAJATEL/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bajatel - Proyecto DAW",
        "description": "API REST para gestión de contratos de telecomunicaciones (Fibra, TV, Móvil)",
        "contact": {
            "email": "soporte@bajatel.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/usuario/registro": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar nuevo usuario",
                "operationId": "43361b0156e412663c4c205817c00d66",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellidos",
                                    "dni",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "García López"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "maxLength": 9,
                                        "minLength": 9,
                                        "example": "12345678A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6,
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/usuario/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión",
                "operationId": "fad3dec8b27517a3a7c3df4e6a4ec299",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso - Devuelve token"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/usuario/eliminarCuenta": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Eliminar cuenta del usuario autenticado",
                "operationId": "226f4c984e21ed9b645fbe9482523e09",
                "responses": {
                    "200": {
                        "description": "Cuenta eliminada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/usuario/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cerrar sesión",
                "operationId": "568f17fb2b8f268397a5d0731a2bb05b",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/usuario/editarPerfil": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Editar perfil del usuario",
                "operationId": "0a834bc086418d2436d1482f6645e1c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "passwordActual"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "passwordActual": {
                                        "type": "string"
                                    },
                                    "passwordNueva": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado"
                    },
                    "401": {
                        "description": "Contraseña actual incorrecta"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/usuarios": {
            "get": {
                "tags": [
                    "Admin - Usuarios"
                ],
                "summary": "Listar todos los usuarios con sus contratos",
                "operationId": "2ec243f8ca47e2227b693f2a8e2c2a0d",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado - Requiere rol gestor/admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/usuario/{id}/gestionarRol": {
            "put": {
                "tags": [
                    "Admin - Usuarios"
                ],
                "summary": "Cambiar rol de usuario (cliente ↔ gestor)",
                "operationId": "520fb0a4e5a14d1b6e38291c5645dd26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol modificado correctamente"
                    },
                    "400": {
                        "description": "No se puede modificar rol de admin"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado - Solo admin"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/usuario/{id}": {
            "delete": {
                "tags": [
                    "Admin - Usuarios"
                ],
                "summary": "Eliminar usuario por ID",
                "operationId": "75f4b508c381e69e781af1fac2bfb9e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado correctamente"
                    },
                    "400": {
                        "description": "No se puede eliminar admin"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado - Solo admin"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum - Usar: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}